profile + JWT

+++ pages/api/post/[id].ts
@@ -1,9 +1,19 @@
 // pages/api/post/[id].ts
 
 import prisma from '../../../lib/prisma';
+import {JWTmiddleware } from '../../../middlewares/JWT_middleware';
+import { getSession } from 'next-auth/react';
+
 
 // DELETE /api/post/:id
 export default async function handle(req, res) {
+  const session = await getSession({ req });
+  if(!JWTmiddleware(req,session))
+  {
+      res.json({"error":"Forbidden"})
+      return;
+  }
+
   const postId = req.query.id;
   if (req.method === 'DELETE') {
     const post = await prisma.post.delete({
